refactor: 회원가입, 로그인 AuthController, AuthService 합치기
