+++ src/main/java/com/business/i4_be/domain/user/dto/response/SigninResponse.java
@@ -1,6 +1,5 @@
 package com.business.i4_be.domain.user.dto.response;
 
-import com.business.i4_be.domain.user.entity.UserRole;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
